fix user_id integer to bigint

+++ app/warehouse_stock/models.py
@@ -1,6 +1,6 @@
 from app.db.database import Base
 from sqlalchemy import Column, Integer, String, DateTime
-
+from sqlalchemy import BigInteger
 
 from sqlalchemy import UniqueConstraint
 
@@ -12,7 +12,7 @@ class WarehouseStocks(Base):
     name = Column(String, nullable=False)
     vid = Column(String)
     brend = Column(String)
-    kod = Column(String)      # ← строка, как ты указал
+    kod = Column(String)
     price = Column(String)
     ostatok = Column(String)
     sklad = Column(String)
@@ -26,6 +26,6 @@ class UserQueryLog(Base):
     __tablename__ = "user_query_log"
 
     id = Column(Integer, primary_key=True)
-    user_id = Column(Integer, nullable=False)
+    user_id = Column(BigInteger, nullable=False)
     query = Column(String, nullable=False)
     timestamp = Column(DateTime, nullable=False)
